+++ resources/assets/lib/wiki-search-suggestions.tsx
@@ -16,59 +16,9 @@
  *    along with osu!web.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { route } from 'laroute';
+import { SuggestionJSON } from 'interfaces/wiki-suggestion-json';
 import * as React from 'react';
-
-export interface SuggestionJSON {
-  html: string;
-  clean: string;
-}
-
-interface SuggestionProps {
-  highlighted: boolean;
-  position: number;
-  suggestion: SuggestionJSON;
-}
-
-class Suggestion extends React.PureComponent<SuggestionProps, {}> {
-  onClick = (e: React.MouseEvent) => {
-    this.publishEvent('select', e);
-  }
-
-  onMouseEnter = (e: React.MouseEvent) => {
-    this.publishEvent('mouseenter', e);
-  }
-
-  onMouseLeave = (e: React.MouseEvent) => {
-    this.publishEvent('mouseleave', e);
-  }
-
-  publishEvent(type: string, e: React.MouseEvent) {
-    const el = e.target as HTMLDivElement;
-
-    if (el.dataset.position != null) {
-      $.publish(`suggestion:${type}`, Number.parseInt(el.dataset.position, 10));
-    }
-  }
-
-  render() {
-    return (
-      <div
-        className={osu.classWithModifiers('wiki-search-suggestions__suggestion', [this.props.highlighted ? 'highlighted' : ''])}
-        data-position={this.props.position}
-        onMouseEnter={this.onMouseEnter}
-        onMouseLeave={this.onMouseLeave}
-        onClick={this.onClick}
-      >
-        <a
-          className='wiki-search-suggestions__suggestion-text'
-          href={route('search', {query: this.props.suggestion.clean, mode: 'wiki_page'})}
-          dangerouslySetInnerHTML={{__html: this.props.suggestion.html}}
-        />
-      </div>
-    );
-  }
-}
+import { Suggestion } from 'wiki-search-suggestion';
 
 interface SuggestionsProps {
   highlighted: number|null;
